Get build working again

+++ foa_mesh/build.rs
@@ -25,7 +25,7 @@ fn main() {
             (
                 "MAX_NUM_PEERS",
                 "Maximum amount of peers connected at the same time.",
-                Value::Integer(5),
+                Value::Integer(4),
                 Some(Validator::PositiveInteger),
             ),
         ],
